requirements.txt
Flask
gevent
python-decouple
python-pixabay
spacy